% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpletasks.R
\name{simpletask}
\alias{simpletask}
\alias{simpletask.default}
\alias{format.simpletask}
\alias{print.simpletask}
\title{simpletask: A Simple Task Object}
\usage{
simpletask(pkg, ver, ...)

\method{simpletask}{default}(pkg, ver)

\method{format}{simpletask}(x, ...)

\method{print}{simpletask}(x, ...)
}
