\name{tinyqueue-package}
\alias{tinyqueue-package}
\docType{package}
\title{\packageTitle{tinyqueue}}
\description{\packageDescription{tinyqueue}}
\section{Package Content}{\packageIndices{tinyqueue}}
\author{\packageAuthor{tinyqueue}}
\section{Maintainer}{\packageMaintainer{tinyqueue}}
\keyword{package}
